packages/xchainpy_crypto/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "xchainpy2_crypto"
version = "0.0.7"
authors = [
    { name = "Tirinox", email = "developer@tirinox.ru" },
]
description = "XChainPy2 Crypto utils and keystore management"
readme = "README.md"
requires-python = ">=3.7"
keywords = ["Crypto", "THORChain", "Blockchain", "XChain"]
license = { text = "MIT" }
urls = { source = "https://github.com/tirinox/xchainpy" }
classifiers = [
    'Development Status :: 3 - Alpha',
    # Chose either "3 - Alpha", "4 - Beta" or "5 - Production/Stable" as the current state of your package
    'Intended Audience :: Developers', # Define that your audience are developers
    'Topic :: Software Development :: Build Tools',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
dependencies = [
    "bip-utils>=2.7.0,<3.0.0",
    "pycryptodome>=3.17,<4.0",
]
[project.optional-dependencies]
test = ["pytest"]

[tool.setuptools]
packages = ["xchainpy2_crypto"]